fuzzer: Put all fuzzing code into own sources
