+++ test/linux/fuzz/service.cpp
@@ -15,19 +15,29 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <cmath> // rand()
 #include <os>
-#include <hw/devices.hpp>
 #include <kernel/events.hpp>
 #include <drivers/usernet.hpp>
-#include <net/inet>
-#include "../router/async_device.hpp"
+#include "macfuzzy.hpp"
 
-static std::unique_ptr<Async_device> dev1;
-static std::unique_ptr<Async_device> dev2;
-static std::vector<uint8_t> load_file(const std::string& file);
-#define htons(x) __builtin_bswap16(x)
+static fuzzy::AsyncDevice dev1;
+static fuzzy::AsyncDevice dev2;
 static const int TCP_PORT = 12345;
+
+std::vector<uint8_t> load_file(const std::string& file)
+{
+	FILE *f = fopen(file.c_str(), "rb");
+	if (f == nullptr) return {};
+	fseek(f, 0, SEEK_END);
+	const size_t size = ftell(f);
+	fseek(f, 0, SEEK_SET);
+  std::vector<uint8_t> data(size);
+	if (size != fread(data.data(), sizeof(char), size, f)) {
+		return {};
+	}
+	fclose(f);
+	return data;
+}
 
 void Service::start()
 {
@@ -70,162 +80,14 @@ void Service::start()
       printf("Writing test to new connection\n");
       connection->write("test");
     });
-}
-
-enum layer_t {
-  ETH,
-  IP4,
-  TCP,
-  UDP,
-  DNS,
-  TCP_CONNECTION
-};
-
-#include "fuzzy_helpers.hpp"
-
-static uint8_t*
-add_ip4_layer(uint8_t* data, FuzzyIterator& fuzzer,
-              const net::ip4::Addr src_addr,
-              const net::ip4::Addr dst_addr,
-              const uint8_t protocol = 0)
-{
-  auto* hdr = new (data) net::ip4::Header();
-  hdr->ttl      = 64;
-  hdr->protocol = (protocol) ? protocol : fuzzer.steal8();
-  hdr->check    = 0;
-  hdr->tot_len  = htons(sizeof(net::ip4::Header) + fuzzer.size);
-  hdr->saddr    = src_addr;
-  hdr->daddr    = dst_addr;
-  //hdr->check    = net::checksum(hdr, sizeof(net::ip4::header));
-  fuzzer.increment_data(sizeof(net::ip4::Header));
-  return &data[sizeof(net::ip4::Header)];
-}
-static uint8_t*
-add_udp4_layer(uint8_t* data, FuzzyIterator& fuzzer,
-              const uint16_t dport)
-{
-  auto* hdr = new (data) net::UDP::header();
-  hdr->sport = htons(fuzzer.steal16());
-  hdr->dport = htons(dport);
-  hdr->length = htons(fuzzer.size);
-  hdr->checksum = 0;
-  fuzzer.increment_data(sizeof(net::UDP::header));
-  return &data[sizeof(net::UDP::header)];
-}
-static uint8_t*
-add_tcp4_layer(uint8_t* data, FuzzyIterator& fuzzer,
-              const uint16_t dport)
-{
-  auto* hdr = new (data) net::tcp::Header();
-  hdr->source_port      = htons(1234);
-  hdr->destination_port = htons(dport);
-  hdr->seq_nr      = fuzzer.steal32();
-  hdr->ack_nr      = fuzzer.steal32();
-  hdr->offset_flags.offset_reserved = 0;
-  hdr->offset_flags.flags = fuzzer.steal8();
-  hdr->window_size = fuzzer.steal16();
-  hdr->checksum    = 0;
-  hdr->urgent      = 0;
-  fuzzer.increment_data(sizeof(net::tcp::Header));
-  return &data[sizeof(net::tcp::Header)];
-}
-static uint8_t*
-add_eth_layer(uint8_t* data, FuzzyIterator& fuzzer, net::Ethertype type)
-{
-  auto* eth = (net::ethernet::Header*) data;
-  eth->set_src({0x1, 0x2, 0x3, 0x4, 0x5, 0x6});
-  eth->set_dest({0x7, 0x8, 0x9, 0xA, 0xB, 0xC});
-  eth->set_type(type);
-  fuzzer.increment_data(sizeof(net::ethernet::Header));
-  return &data[sizeof(net::ethernet::Header)];
-}
-
-static inline uint16_t udp_port_scan(net::Inet& inet)
-{
-  for (uint16_t udp_port = 1; udp_port < 65535; udp_port++) {
-    if (inet.udp().is_bound({inet.ip_addr(), udp_port})) {
-      return udp_port;
-    }
-  }
-  return 0;
-}
-
-static void
-insert_into_stack(layer_t layer, const uint8_t* data, const size_t size)
-{
-  auto& inet = net::Super_stack::get(0);
-  const size_t packet_size = std::min((size_t) inet.MTU(), size);
-  FuzzyIterator fuzzer{data, packet_size};
   
-  auto p = inet.create_packet();
-  // link layer -> IP4
-  auto* eth_end = add_eth_layer(p->layer_begin(), fuzzer,
-                                net::Ethertype::IP4);
-  // select layer to fuzz
-  switch (layer) {
-  case ETH:
-    // by subtracting 2 i can fuzz ethertype as well
-    fuzzer.fill_remaining(eth_end);
-    break;
-  case IP4:
-    {
-      auto* next_layer = add_ip4_layer(eth_end, fuzzer,
-                         {10, 0, 0, 1}, inet.ip_addr());
-      fuzzer.fill_remaining(next_layer);
-      break;
-    }
-  case UDP:
-    {
-      // scan for UDP port (once)
-      static uint16_t udp_port = 0;
-      if (udp_port == 0) {
-        udp_port = udp_port_scan(inet);
-        assert(udp_port != 0);
-      }
-      // generate IP4 and UDP datagrams
-      auto* ip_layer = add_ip4_layer(eth_end, fuzzer,
-                         {10, 0, 0, 1}, inet.ip_addr(),
-                          (uint8_t) net::Protocol::UDP);
-      auto* udp_layer = add_udp4_layer(ip_layer, fuzzer,
-                          udp_port);
-      fuzzer.fill_remaining(udp_layer);
-      break;
-    }
-  case TCP:
-    {
-      // generate IP4 and TCP data
-      auto* ip_layer = add_ip4_layer(eth_end, fuzzer,
-                         {10, 0, 0, 1}, inet.ip_addr(),
-                          (uint8_t) net::Protocol::TCP);
-      auto* tcp_layer = add_tcp4_layer(ip_layer, fuzzer,
-                          TCP_PORT);
-      fuzzer.fill_remaining(tcp_layer);
-      break;
-    }
-  case TCP_CONNECTION:
-    //
-    break;
-  default:
-    assert(0 && "Implement me");
-  }
-  // we have to add ethernet size here as its not part of MTU
-  p->set_data_end(fuzzer.data_counter);
-  dev1->get_driver()->receive(std::move(p));
-}
-
-std::vector<uint8_t> load_file(const std::string& file)
-{
-	FILE *f = fopen(file.c_str(), "rb");
-	if (f == nullptr) return {};
-	fseek(f, 0, SEEK_END);
-	const size_t size = ftell(f);
-	fseek(f, 0, SEEK_SET);
-  std::vector<uint8_t> data(size);
-	if (size != fread(data.data(), sizeof(char), size, f)) {
-		return {};
-	}
-	fclose(f);
-	return data;
+  printf(R"FUzzY(
+   __  __                     ___                            _  _
+  |  \/  |  __ _     __      | __|  _  _      ___     ___   | || |
+  | |\/| | / _` |   / _|     | _|  | +| |    |_ /    |_ /    \_, |
+  |_|__|_| \__,_|   \__|_   _|_|_   \_,_|   _/__|   _/__|   _|__/
+  _|"""""|_|"""""|_|"""""|_| """ |_|"""""|_|"""""|_|"""""|_| """"|
+  "`-0-0-'"`-0-0-'"`-0-0-'"`-0-0-'"`-0-0-'"`-0-0-'"`-0-0-'"`-0-0-' )FUzzY");
 }
 
 // libfuzzer input
@@ -240,6 +102,14 @@ int LLVMFuzzerTestOneInput(const uint8_t* data, size_t size)
     userspace_main(1, args);
   }
   if (size == 0) return 0;
-  insert_into_stack(UDP, data, size);
+  
+  // IP-stack fuzzing
+  fuzzy::stack_config config {
+    .layer = fuzzy::IP4,
+    .ip_port = 0
+  };
+  fuzzy::insert_into_stack(dev1, config, data, size);
+  // Upper layer fuzzing using fuzzy::Stream
+  
   return 0;
 }
